Move functions around and remove includes to reduce dependencies in runtime configuration
