+++ src/runtime/threads/resource_manager.cpp
@@ -7,7 +7,6 @@
 #include <hpx/config.hpp>
 #include <hpx/assertion.hpp>
 #include <hpx/error_code.hpp>
-#include <hpx/lcos/local/once.hpp>
 #include <hpx/runtime/threads/resource_manager.hpp>
 #include <hpx/runtime/threads/thread_executor.hpp>
 #include <hpx/util/reinitializable_static.hpp>
